Revert "Makes jquery-ui an external dependency, taking it out of the repo"
This reverts commit e966f57a1955daa6e97bad303c94ef89a9255f9d.

jQuery-UI API has changed in an incompatible way after 1.8.9. Updating our code
now does not guarantee this won't happend again in the future. Hence, the safest
solution is to revert the patch that made jQuery-UI an external dependency.

Also, since in the near future the UI will migrate from jQuery-UI to a different
framework, it is ok to keep all this code in the repo by now.